+++ apps/api/src/websocket/websocket.module.ts
@@ -1,5 +1,6 @@
 import { Global, Module } from "@nestjs/common";
 
+import { ChessModule } from "src/chess/chess.module";
 import { WsJwtGuard } from "src/websocket/guards/ws-jwt.guard";
 import { REALTIME_PUBLISHER } from "src/websocket/realtime.publisher";
 import { SocketRealtimePublisher } from "src/websocket/realtime.publisher.service";
@@ -7,6 +8,7 @@ import { WsGateway } from "src/websocket/websocket.gateway";
 
 @Global()
 @Module({
+  imports: [ChessModule],
   providers: [
     WsGateway,
     WsJwtGuard,